Rewrite source map writer
Users may set permalink for compileOption and the previous code did not
address that.

This plugin writes source map files at the same directory of its CSS
file and the filename is the CSS filename + ".map".

To write source map files at the right directory, add a transform which
does not transform contents, but writes source map files, because output
paths for CSSes generated from Sass/SCSS files are available in
transforms.