Fix watcher on Windows
Issue: #12
Watcher not recompiling on changes on Windows

The bug was caused by an inconsistency of paths in Eleventy.
The inputPath argument of compile(), getCacheKey(), etc. is the posix
style, but the argument of "eleventy.beforeWatch" event is an array of
paths that are a platform-dependent style.

Therefore, dependencyMap should store dependants as the posix style, but
dependencies as a platform-dependent style.