Fix breaks with eleventyComputed.js #13
Issue: #13
Breaks with eleventyComputed.js

If a property in computed data has a string value, the string will be
resolved with other data properties.
https://www.11ty.dev/docs/data-computed/#using-a-template-string
https://github.com/11ty/eleventy/blob/0bbc16dfc03eb773de33488dd1c7b1bb37f701be/src/Template.js#L550-L558

If you have a property with a String value in _data/eleventyComputed.js,
all of the templates' compile() functions are called with the string
value as inputContent, since _data/eleventyComputed.js is global.

The problem is Sass/SCSS compiler does not accept arbitrary string
values. The string should be Sass/SCSS format.

Therefore, the plugin's compile function need to know if it is a file
content or a value of a property in eleventyComputed.

However, it is impossible, because Eleventy does not give such
information to any templates other than Markdown.
https://github.com/11ty/eleventy/blob/0bbc16dfc03eb773de33488dd1c7b1bb37f701be/src/TemplateRender.js#L276-L293

The fix is essentially patching Eleventy code to provide the above
missing information to the plugin's compile() function and making the
function use the information to avoid compiling property values.