Fix a compatibility issue against 2.0.0-canary.18
In @11ty/eleventy@2.0.0-canary.18, a dependency monitoring issue was
fixed.  ( 11ty/eleventy#2479 )
The fix is OK, but Eleventy had been using platform dependent paths only
in watchers before 2.0.0-canary.18, and now it uses posix paths all of
the places.
As a result, the argument of beforeWatch events was changed from
platform dependent paths to posix paths. It is a breaking API change and
the cause of this compatibility issue.

By this commit, eleventy-sass also uses the fix before updating
dependencyMap, if it detects the fix (PathNormalizer) in Eleventy.