Fix a compatibility issue with 2.0.0-canary.24
By this commit, you will be able to compile Sass/SCSS files in an
Eleventy 2.0.0-canary.24 and above project, although the result CSS
files might be wrong, because there is still an unresolved issue.

The unresolved issue (with @11ty/eleventy-2.0.0-canary.19 and above):
If you have a property with a string value in an eleventyComputed in any
data files that affects a Sass/SCSS file, the result CSS files will be
wrong.