Add warning for missing necessary option
The plugin uses an experimental feature of Node.js 22 to load Eleventy 3
internal modules. Users must add the `--experimental-require-module`
option to their Node commands.

If loading an Eleventy 3 internal module fails, it is likely due to the
absence of the `--experimental-require-module` option. The plugin will
output a warning message to notify the user of the necessity of adding
this option to their command.